requirements.txt
fastapi
uvicorn[standard]

pandas
numpy
python-dotenv

# ---- LangChain (SAFE versions for Google GenAI + Python 3.13) ----
langchain>=0.1.8
langchain-core>=0.1.8
langchain-community
langgraph
langchain-google-genai>=0.0.6
google-generativeai

# ---- Vector Store ----
sentence-transformers
chromadb

# ---- File uploads ----
python-multipart

# ---- Visualization ----
plotly
matplotlib
seaborn

# ---- Pydantic v2 (needed for FastAPI & LangChain core) ----
pydantic>=2.0

# ---- Server stability & async ----
aiohttp

# visualization & frontend
streamlit
kaleido